img-search/requirements.txt
chromadb
streamlit
Pillow
torch
transformers
sentence-transformers
mysql-connector-python
tidb-vector[client]
pymysql
mysql
python-dotenv